Better handling of unit prefixes, renamed readFromString
Unit prefixes are now handled in a better way. Eventually,
consolidating Quants to use a single prefix value is in order (i.e., do
away with tuples and just use a list of powers and a single prefix
value for units). Only a single prefix value is needed in nearly all
cases---it's just a stand-in for "times ten to the nth power."

readFromString was renamed as "fromStringToQuant."